414, 468, 486, 624, 668
parse_args prints to stderr a usage error and the error message and then
throws a SystemExit exception. Specs say if arguments are contradictory
(which I thought meant was contradictory with itself but still valid
invocations, but which was judged to include invalid invocations as well)
then we must only print "invalid" and return -1 rather than allowing python
to print it's friendly error message. Temporarily reroute stderr for parse_args
so that the error message doesn't go through and catch the exception so that
we can follow the spec more precisely