643, 644
System Exit is thrown on multiple argument.  Did not know arguments
were not supposed to be specified twice so wrote a helper function
to dig into and replace the arg parse library because apparently
it is not valid to simply use the last value of the repeated
invocation of an argument and should instead throw an invalid/exit.